LibWeb: Move table grid slots functionality to separate module
